remove now-unused modules
