+++ nexus/reconfigurator/planning/src/blueprint_builder/mod.rs
@@ -6,8 +6,6 @@
 
 mod builder;
 mod clickhouse;
-mod external_networking;
-mod internal_dns;
 
 pub use builder::*;
 pub use clickhouse::{ClickhouseAllocator, ClickhouseZonesThatShouldBeRunning};
