+++ nexus/reconfigurator/planning/src/blueprint_editor.rs
@@ -28,3 +28,6 @@ pub(crate) use allocators::ExternalNetworkingChoice;
 pub(crate) use allocators::ExternalSnatNetworkingChoice;
 pub(crate) use sled_editor::EditedSled;
 pub(crate) use sled_editor::SledEditor;
+
+#[cfg(test)]
+pub(crate) use allocators::ExternalIpAllocator;
